+++ RELEASE_NOTES.md
@@ -3,6 +3,10 @@ Release Notes
 ##vNext
 * Dashboards: Initial support for creating dashboards
 
+## 1.6.10
+* Azure SQL Server: geo_replicate parameter to geo-replicate the server databases
+* App Insights: Support for Availability tests, VS WebTests
+
 ## 1.6.9
 * Resource Groups: Support for creating resource groups for deployments targeting a subscription.
 * WebApp: Slots now inherit user assigned identities from their owning webApp